Removed rest package dependency by using json instead.
Previously, rest::format_json was used to turn a JSON string into a Tcl dict.
However, the rest package required a package called 'tdom', which I did not
obtain by installing tcllib.

The tdom package had to be downloaded and built from source. The process is
straightforward, but not worth the extra effort. Besides, the rest package
(version 1.0.2 as of writing) seemed a bit clunky and outdated.