Decoupled sessions from gateway and updated discord::gateway::connect arguments.
The concept of sessions in the discord namespace is removed from the gateway so
that there is freedom to implement a layer around the gateway.

The connect procedure now accepts a list for the cmd argument instead of just a
name. This allows additional arguments to be placed in front of the last
argument for the WebSocket object.