Used local HTTPS server for test GetGateway-1.4.
The problem with TLS sockets seems to be a deadlock if you try to run a server
and client in the same thread. Might want to investigate further, or contact
the maintainers.

In any case HTTP will suffice for now, and there will be future tests to make
sure the content sent is the same regardless of TLS being present.